+++ Source/shared/TraktToken.swift
@@ -9,18 +9,17 @@
 import Foundation
 
 /// Represents a trakt token
-public class TraktToken {
-    private static var userDefaultsTokenKey = "traktToken"
+@objc public class TraktToken: NSObject, NSCoding {
     /// The access token
-    public let accessToken: String!
+    public let accessToken: String
     /// Expiration date
-    public let expiresAt: NSDate!
+    public let expiresAt: NSDate
     /// Refresh value
-    public let refreshToken: String!
+    public let refreshToken: String
     /// Token's type
-    public let tokenType: String!
+    public let tokenType: String
     /// Token's scope
-    public let scope: String!
+    public let scope: String
 
     /**
      Init a token
@@ -31,7 +30,7 @@ public class TraktToken {
      - parameter tokenType:    tokenType
      - parameter scope:        scope
      */
-    init(accessToken: String, expiresAt: NSDate, refreshToken: String, tokenType: String, scope: String) {
+    public init(accessToken: String, expiresAt: NSDate, refreshToken: String, tokenType: String, scope: String) {
         self.accessToken = accessToken
         self.expiresAt = expiresAt
         self.refreshToken = refreshToken
@@ -40,62 +39,49 @@ public class TraktToken {
     }
 
     /**
-     Init with data
+     Init with a JSONHash
+
      - parameter data: JSONHash
      */
-    convenience init?(data: JSONHash!) {
-        guard let token = data,
-            accessToken = token["access_token"] as? String,
-            expiresIn = token["expires_in"] as? Double,
-            scope = token["scope"] as? String,
-            tokenType = token["token_type"] as? String,
-            refreshToken = token["refresh_token"] as? String else {
+    convenience init?(data: JSONHash?) {
+        guard let accessToken = data?["access_token"] as? String,
+            expiresIn = data?["expires_in"] as? Double,
+            scope = data?["scope"] as? String,
+            tokenType = data?["token_type"] as? String,
+            refreshToken = data?["refresh_token"] as? String else {
                 return nil
         }
         self.init(accessToken: accessToken, expiresAt: NSDate(timeIntervalSinceNow: expiresIn), refreshToken: refreshToken, tokenType: tokenType, scope: scope)
     }
 
-    /**
-     Attemps to load token from NSUserDefaults
-
-     - parameter clientId: clientId
-
-     - returns: TraktToken
-     */
-    internal static func load(clientId: String) -> TraktToken? {
-        let defaults = NSUserDefaults.standardUserDefaults()
-        guard let data = defaults.objectForKey(TraktToken.userDefaultsTokenKey + clientId) as? JSONHash else {
-            return nil
+    /// NSCoding
+    public required init?(coder aDecoder: NSCoder) {
+        guard let accessToken = aDecoder.decodeObjectForKey("accessToken") as? String,
+        tokenType = aDecoder.decodeObjectForKey("tokenType") as? String,
+        expiresAt = aDecoder.decodeObjectForKey("expiresAt") as? NSDate,
+        refreshToken = aDecoder.decodeObjectForKey("refreshToken") as? String,
+            scope = aDecoder.decodeObjectForKey("scope") as? String else {
+                return nil
         }
-        return TraktToken(data: data)
+
+        print(accessToken)
+        print(expiresAt)
+        print(refreshToken)
+        print(tokenType)
+        print(scope)
+        self.accessToken = accessToken
+        self.expiresAt = expiresAt
+        self.refreshToken = refreshToken
+        self.tokenType = tokenType
+        self.scope = scope
     }
 
-    /**
-     Save token to NSUserDefaults
-
-     - parameter clientId: clientId
-     */
-    internal func save(clientId: String!) {
-        NSUserDefaults.standardUserDefaults().setObject(data, forKey: TraktToken.userDefaultsTokenKey + clientId)
-    }
-
-    /**
-     Remove token from NSUserDefaults
-
-     - parameter clientId: clientId
-     */
-    internal func remove(clientId: String!) {
-        NSUserDefaults.standardUserDefaults().removeObjectForKey(TraktToken.userDefaultsTokenKey + clientId)
-    }
-
-    /// Give a JSONHash represenation of the token
-    private var data: JSONHash {
-        return [
-            "access_token": accessToken,
-            "token_type": tokenType,
-            "expires_in": expiresAt.timeIntervalSinceNow,
-            "refresh_token": refreshToken,
-            "scope": scope
-        ]
+    /// NSCoding
+    public func encodeWithCoder(aCoder: NSCoder) {
+        aCoder.encodeObject(accessToken, forKey: "accessToken")
+        aCoder.encodeObject(tokenType, forKey: "tokenType")
+        aCoder.encodeObject(expiresAt, forKey: "expiresAt")
+        aCoder.encodeObject(refreshToken, forKey: "refreshToken")
+        aCoder.encodeObject(scope, forKey: "scope")
     }
 }