+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log
 All notable changes to this project will be documented in this file.
-After the revision 1.0.0
+
+---
+
+# 1.0.2
+
+### Added
+- Code documentation
+- Logged tests
+- Refresh Token request
+
+### Updated
+- Rewrited Requests use generics: add/remove history/watchlist, add rating
+- TrakToken conforms to NSObject, NSCoding
+
+### Removed
+- TraktType enum
 
 ---
 
@@ -8,10 +23,12 @@ After the revision 1.0.0
 
 ### Added
 - Code documentation
-- Allow TraktObject to be extended with another one of the same type
+- Allow TraktObject to be extended the same object type
+
 ### Fixed
 - #8 Empty seasons in TraktShow
-#### Updated
+
+### Updated
 - enum TraktMediaType removed use TraktType instead
 
 ---